Creates separate rankings.py folder. About to separate Pydantic models into separate file.
